+++ app.js
@@ -1,3 +1,5 @@
+'use strict';
+
 var commandManager = require('./modules/command-manager');
 var telegramBot = require('./modules/telegram-bot');
 var genius = require('./modules/genius');
