+++ bin/bot
@@ -1,4 +1,5 @@
 #!/usr/bin/env node
+'use strict';
 
 // catch the uncaught errors that weren't wrapped in a domain or try catch statement
 // do not use this in modules, but only in applications, as otherwise we could have multiple of these bound
