+++ modules/command-manager/index.js
@@ -1,3 +1,5 @@
+'use strict';
+
 /**
  * CommandManager take a message, do an action and call a callback for response
  * @constructor
