+++ modules/dipartimenti.js
@@ -1,3 +1,5 @@
+'use strict';
+
 module.exports = [
     {
         id: 0,
