+++ modules/speaker/index.js
@@ -1,3 +1,5 @@
+'use strict';
+
 /**
  * This module ask a question to the user resolving a Promise when the user answers.
  * @module
